+++ src/biz/bokhorst/xprivacy/PackageChange.java
@@ -53,7 +53,7 @@ public class PackageChange extends BroadcastReceiver {
 
 							// Restrict new non-system apps
 							if (!appInfo.isSystem())
-								PrivacyManager.applyTemplate(uid, null, true);
+								PrivacyManager.applyTemplate(uid, null, true, true);
 						}
 					}